@@ -11,21 +11,16 @@
         }
     ],
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.4",
         "simplesamlphp/composer-module-installer": "~1.1"
     },
     "require-dev": {
-        "simplesamlphp/simplesamlphp": ">=1.17",
-        "phpunit/phpunit": "~5.7"
+        "simplesamlphp/simplesamlphp": ">=1.19",
+        "simplesamlphp/simplesamlphp-test-framework": "~0.1.0"
     },
     "autoload": {
         "psr-4": {
             "SimpleSAML\\Module\\genuniqueid\\": "lib/"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "SimpleSAML\\Test\\Utils\\": "vendor/simplesamlphp/simplesamlphp/tests/Utils"
         }
     },
     "support": {
