@@ -1,9 +1,9 @@
 <?php
 $projectRoot = dirname(__DIR__);
-require_once($projectRoot.'/vendor/autoload.php');
+require_once($projectRoot . '/vendor/autoload.php');
 
 // Symlink module into ssp vendor lib so that templates and urls can resolve correctly
-$linkPath = $projectRoot.'/vendor/simplesamlphp/simplesamlphp/modules/consent';
+$linkPath = $projectRoot . '/vendor/simplesamlphp/simplesamlphp/modules/consent';
 if (file_exists($linkPath) === false) {
     echo "Linking '$linkPath' to '$projectRoot'\n";
     symlink($projectRoot, $linkPath);
